Fix sync backup

+++ internal/server/config.go
@@ -7,15 +7,15 @@ import (
 )
 
 type Config struct {
-	Addr            string  `env:"ADDRESS"`
-	StoreInterval   float64 `env:"STORE_INTERVAL"`
-	FileStoragePath string  `env:"FILE_STORAGE_PATH"`
-	Restore         bool    `env:"RESTORE"`
+	Addr            string `env:"ADDRESS"`
+	StoreInterval   uint64 `env:"STORE_INTERVAL"`
+	FileStoragePath string `env:"FILE_STORAGE_PATH"`
+	Restore         bool   `env:"RESTORE"`
 }
 
 func ParseConfig() (*Config, error) {
 	addr := flag.String("a", "localhost:8080", "Адрес сервера (в формате host:port)")
-	storeInterval := flag.Float64("i", 300, "Интервал сохранения данных на диск (в секундах)")
+	storeInterval := flag.Uint64("i", 300, "Интервал сохранения данных на диск (в секундах)")
 	fileStoragePath := flag.String("f", "/tmp/.monit/memstorage.json", "Путь до файла сохранения данных")
 	restore := flag.Bool("r", true, "Флаг загрузки сохраненных данных при старте сервера")
